requirements.txt
# Core ML and Data Science Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Modern NLP Libraries
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
tokenizers>=0.13.0
accelerate>=0.20.0

# Text Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Web UI Framework
streamlit>=1.25.0
plotly>=5.15.0

# Database and Storage
sqlite3
sqlalchemy>=2.0.0

# Additional Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
